server/requirements.txt
Flask
openai
waitress
flask-cors
python-dotenv